create LanguagePolicy declarative object with .schema files Create MultiLanguageRecognizer to allow you to define multiple recognizers per language with language policy Added unit tests
