+++ libraries/Microsoft.Bot.Builder.Dialogs/ILanguagePolicy.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Bot.Builder.Dialogs
+{
+    /// <summary>
+    /// Language policy map
+    /// "en-us" -> "en-us","en",""
+    /// </summary>
+    public interface ILanguagePolicy : IDictionary<string, string[]>
+    {
+    }
+}
