+++ libraries/Microsoft.Bot.Builder.Dialogs/Schemas/Policies/Microsoft.LanguagePolicy.schema
@@ -1,0 +1,14 @@
+{
+    "$schema": "https://raw.githubusercontent.com/Microsoft/botbuilder-tools/SchemaGen/packages/CogSchema/src/cogSchema.schema",
+    "$role": "unionType(Microsoft.ILanguagePolicy)",
+    "title": "Language Policy",
+    "description": "This represents a dialog which gathers a DateTime in a specified range",
+    "type": "object",
+    "additionalProperties": {
+        "type": "array",
+        "items": {
+            "type": "string",
+            "title": "language code like en-us"
+        }
+    }
+}
